requirements.txt
asgiref==3.5.2
cfgv==3.3.1
distlib==0.3.6
Django==4.1.3
django-throttle-requests==0.6.0
filelock==3.8.0
identify==2.5.9
mccabe==0.7.0
nodeenv==1.7.0
numpy==1.23.5
platformdirs==2.5.4
pre-commit==2.20.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
PyYAML==6.0
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.6
virtualenv==20.16.7
